tweak: Reorganize the detour panels into a folder
