+++ assets/src/components/detours/diversionPage.tsx
@@ -6,7 +6,7 @@ import React, {
   useEffect,
   useState,
 } from "react"
-import { DrawDetourPanel } from "./drawDetourPanel"
+import { DrawDetourPanel } from "./detourPanels/drawDetourPanel"
 import { DetourMap } from "./detourMap"
 import { useDetour } from "../../hooks/useDetour"
 import { Alert, CloseButton } from "react-bootstrap"
@@ -14,14 +14,14 @@ import * as BsIcons from "../../helpers/bsIcons"
 import { OriginalRoute } from "../../models/detour"
 import { joinClasses } from "../../helpers/dom"
 import { AsProp } from "react-bootstrap/esm/helpers"
-import { DetourFinishedPanel } from "./detourFinishedPanel"
-import { DetourRouteSelectionPanel } from "./detourRouteSelectionPanel"
+import { DetourFinishedPanel } from "./detourPanels/detourFinishedPanel"
+import { DetourRouteSelectionPanel } from "./detourPanels/detourRouteSelectionPanel"
 import { Route, RoutePattern } from "../../schedule"
 import RoutesContext from "../../contexts/routesContext"
 import { Snapshot } from "xstate"
 import inTestGroup, { TestGroups } from "../../userInTestGroup"
-import { ActiveDetourPanel } from "./activeDetourPanel"
-import { PastDetourPanel } from "./pastDetourPanel"
+import { ActiveDetourPanel } from "./detourPanels/activeDetourPanel"
+import { PastDetourPanel } from "./detourPanels/pastDetourPanel"
 import userInTestGroup from "../../userInTestGroup"
 import { useCurrentTimeSeconds } from "../../hooks/useCurrentTime"
 import { timeAgoLabel } from "../../util/dateTime"
